feat: 全新AT指令实现MQTT、HTTP、OTA等功能

+++ projects/athena-common-embedded/device_info/device_info.h
@@ -12,7 +12,7 @@ void init_temperature();
 
 float get_temperature();
 
-int get_electricity();
+float get_electricity();
 
 
 #endif
